test: converting the TaskDetailFragment and xml files
