+++ app/src/main/java/com/applesforryuk/mylistmaking/DetailActivity.kt
@@ -12,10 +12,10 @@ import com.applesforryuk.mylistmaking.databinding.ActivityDetailBinding
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 
 
-class DetailActivity : AppCompatActivity() {
+class DetailActivity : AppCompatActivity() { /*
     private lateinit var binding: ActivityDetailBinding
     private var list: TaskList? = null
-    lateinit var addTaskButton: FloatingActionButton
+
 
     private var adapter: TaskListAdapter? = null
 
@@ -27,43 +27,11 @@ class DetailActivity : AppCompatActivity() {
         list = intent.getParcelableExtra(MainActivity.INTENT_LIST_KEY) as? TaskList
         title = list?.name
 
-        binding.taskListRecyclerview.layoutManager = LinearLayoutManager(this)
-        adapter = list?.let { TaskListAdapter(it) }
-        binding.taskListRecyclerview.adapter = adapter
 
-        addTaskButton = findViewById(R.id.add_task_button)
-        addTaskButton.setOnClickListener{
-            showCreateTaskDialog()
-        }
+        } */
     }
 
-    @Deprecated("Deprecated in Java")
-    override fun onBackPressed() {
-        val bundle = Bundle()
-        bundle.putParcelable(MainActivity.INTENT_LIST_KEY, list)
-
-        val intent = Intent()
-        intent.putExtras(bundle)
-        setResult(Activity.RESULT_OK, intent)
 
 
-        super.onBackPressed()
-    }
 
-    private fun showCreateTaskDialog() {
-        val taskEditText = EditText(this)
-        taskEditText.inputType = InputType.TYPE_CLASS_TEXT
-        AlertDialog.Builder(this)
-            .setTitle(R.string.task_to_add)
-            .setView(taskEditText)
-            .setPositiveButton(R.string.add_task) {
-                    dialog, _ ->
-                val task = taskEditText.text.toString()
-                list?.tasks?.add(task)
-                dialog.dismiss()
-            }
-            .create()
-            .show()
 
-    }
-}
